fixed a buncha' routing stuff with Kathy
